feat: add dagger, update retrofit and update some models

+++ app/src/main/java/com/example/melitest/data/model/ClaimsModel.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class ClaimsModel(
     @SerializedName("period") var period: String? = null,
     @SerializedName("rate") var rate: Double? = null,
-    @SerializedName("value") var value: Int? = null
+    @SerializedName("value") var value: Long? = null
 )
