+++ app/src/main/java/com/example/melitest/data/model/PagingModel.kt
@@ -3,8 +3,8 @@ package com.example.melitest.data.model
 import com.google.gson.annotations.SerializedName
 
 data class PagingModel(
-    @SerializedName("total") var total: Int? = null,
-    @SerializedName("primary_results") var primaryResults: Int? = null,
-    @SerializedName("offset") var offset: Int? = null,
-    @SerializedName("limit") var limit: Int? = null
+    @SerializedName("total") var total: Long? = null,
+    @SerializedName("primary_results") var primaryResults: Long? = null,
+    @SerializedName("offset") var offset: Long? = null,
+    @SerializedName("limit") var limit: Long? = null
 )
